requirements.txt
ansicolors==1.1.8
sympy